refactor: component 세분화 완료
