+++ lib/presentation/screen/component/bottom_sheet_header.dart
@@ -1,0 +1,34 @@
+import 'package:calendar_scheduler/presentation/extension.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+import '../../const/strings.dart';
+import '../provider/schedule_provider.dart';
+
+class BottomSheetHeader extends StatelessWidget {
+  const BottomSheetHeader({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    String title = context
+        .read<ScheduleProvider>()
+        .currentDateTime
+        .toFormattedString(Strings.DATE_FORMAT);
+    return Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
+      Text(
+        title,
+        style: const TextStyle(
+          fontSize: 16,
+          fontWeight: FontWeight.w800,
+        ),
+      ),
+      IconButton(
+        onPressed: () {
+          Navigator.of(context).pop();
+        },
+        icon: const Icon(Icons.close),
+      ),
+    ]);
+  }
+}
