+++ lib/presentation/screen/component/field/content_input_field.dart
@@ -3,12 +3,15 @@ import 'package:flutter/material.dart';
 import '../../../const/colors.dart';
 import '../../../const/strings.dart';
 import '../../../const/styles.dart';
-import '../../schedule_bottom_sheet/create_schedule_bottom_sheet.dart';
 
 class ContentInputField extends StatefulWidget {
   final String initialContent;
   final ContentSetter contentSetter;
-  ContentInputField({super.key, required this.contentSetter, required this.initialContent});
+  const ContentInputField({
+    super.key,
+    required this.contentSetter,
+    required this.initialContent,
+  });
 
   @override
   State<ContentInputField> createState() => _ContentInputFieldState();
@@ -67,3 +70,5 @@ class _ContentInputFieldState extends State<ContentInputField> {
     super.dispose();
   }
 }
+
+typedef ContentSetter = void Function(String content);
