+++ lib/presentation/screen/component/schedule_bottom_sheet/create_schedule_bottom_sheet.dart
@@ -1,16 +1,14 @@
 import 'package:calendar_scheduler/presentation/const/colors.dart';
-import 'package:calendar_scheduler/presentation/extension.dart';
-import 'package:calendar_scheduler/presentation/screen/component/field/time_input_field.dart';
+import 'package:calendar_scheduler/presentation/screen/component/bottom_sheet_header.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../../const/strings.dart';
-import '../../const/styles.dart';
-import '../component/field/color_selection_field.dart';
-import '../component/field/content_input_field.dart';
-import '../provider/schedule_provider.dart';
-
-
+import '../../../const/strings.dart';
+import '../../../const/styles.dart';
+import '../../provider/schedule_provider.dart';
+import '../field/color_selection_field.dart';
+import '../field/content_input_field.dart';
+import '../field/start_end_time_input_field.dart';
 
 class CreateScheduleBottomSheet extends StatefulWidget {
   final DateTime selectedDate;
@@ -50,7 +48,7 @@ class _CreateScheduleBottomSheetState extends State<CreateScheduleBottomSheet> {
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            const _BottomSheetHeader(),
+            const BottomSheetHeader(),
             ColorSelectionField(
               colors: colors,
               selectedColorId:
@@ -62,7 +60,7 @@ class _CreateScheduleBottomSheetState extends State<CreateScheduleBottomSheet> {
               },
             ),
             const Spacer(),
-            _TimeInputRenderer(),
+            const StartEndTimeInputField(),
             const Spacer(),
             ContentInputField(
               initialContent: context.watch<ScheduleProvider>().currentContent,
@@ -87,68 +85,6 @@ class _CreateScheduleBottomSheetState extends State<CreateScheduleBottomSheet> {
     context
         .watch<ScheduleProvider>()
         .updateCurrentDateTime(widget.selectedDate);
-  }
-}
-
-typedef ContentSetter = void Function(String content);
-typedef ColorIdSetter = void Function(int id);
-typedef TimeSetter = void Function(int time);
-
-class _BottomSheetHeader extends StatelessWidget {
-  const _BottomSheetHeader();
-
-  @override
-  Widget build(BuildContext context) {
-    String title = context
-        .read<ScheduleProvider>()
-        .currentDateTime
-        .toFormattedString(Strings.DATE_FORMAT);
-    return Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
-      Text(
-        title,
-        style: const TextStyle(
-          fontSize: 16,
-          fontWeight: FontWeight.w800,
-        ),
-      ),
-      IconButton(
-        onPressed: () {
-          Navigator.of(context).pop();
-        },
-        icon: const Icon(Icons.close),
-      ),
-    ]);
-  }
-}
-
-class _TimeInputRenderer extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Expanded(
-          child: TimeInputField(
-            initialTime: context.watch<ScheduleProvider>().currentStartTime,
-            selectedTimeType: Strings.LABEL_START_TIME,
-            timeSetter: (int time) {
-              context.read<ScheduleProvider>().updateCurrentStatTime(time);
-            },
-          ),
-        ),
-        const SizedBox(
-          width: 4.0,
-        ),
-        Expanded(
-          child: TimeInputField(
-            initialTime: context.watch<ScheduleProvider>().currentEndTime,
-            selectedTimeType: Strings.LABEL_END_TIME,
-            timeSetter: (int time) {
-              context.read<ScheduleProvider>().updateCurrentEndTime(time);
-            },
-          ),
-        )
-      ],
-    );
   }
 }
